refactor(plugins) uncategorise plugins
