+++ lua/plugins/autopairs.lua
@@ -1,0 +1,5 @@
+local rule = require('nvim-autopairs.rule')
+local pairs = require('nvim-autopairs')
+
+pairs.setup({})
+pairs.add_rule(rule('*', '*', 'markdown'))
