--- lua/plugins/behavior/autopairs.lua
@@ -1,5 +1,0 @@
-local rule = require('nvim-autopairs.rule')
-local pairs = require('nvim-autopairs')
-
-pairs.setup({})
-pairs.add_rule(rule('*', '*', 'markdown'))
