+++ lua/plugins/indentline.lua
@@ -1,0 +1,2 @@
+-- see syntax.lua for IndentMarker
+require("ibl").setup({ indent = { highlight = { 'IndentMarker' } , char = '│' }})
