+++ lua/plugins/init.lua
@@ -49,25 +49,25 @@ local which = require('plugins.whichkey')
 local register = which.register
 
 -- plugin setups, sorted alphabetically
--- require('plugins.external.dap')(register, which.dap)
--- require('plugins.external.dapui')(register, which.dapui)
-require('plugins.external.lsp')(register, which.attach)
-require('plugins.external.null')(register, which.attach)
+-- require('plugins.dap')(register, which.dap)
+-- require('plugins.dapui')(register, which.dapui)
+require('plugins.lsp')(register, which.attach)
+require('plugins.null')(register, which.attach)
 require('leap').add_default_mappings()
 
-require('plugins.behavior.autopairs')
-require('plugins.behavior.cmp')(which.cmp)
-require('plugins.behavior.comment')
-require('plugins.behavior.luasnip')
-require('plugins.behavior.surround')
-require('plugins.behavior.telescope')(register, which.telescope)
+require('plugins.autopairs')
+require('plugins.cmp')(which.cmp)
+require('plugins.comment')
+require('plugins.luasnip')
+require('plugins.surround')
+require('plugins.telescope')(register, which.telescope)
 
-require('plugins.interface.bufferline')
-require('plugins.interface.ccc')
-require('plugins.interface.gitsigns')
-require('plugins.interface.indentline')
-require('plugins.interface.lualine')
-require('plugins.interface.nvim-tree')(register, which.nvimtree)
-require('plugins.interface.scrollbar')
-require('plugins.interface.treesitter')
-require('plugins.interface.virtcolumn')
+require('plugins.bufferline')
+require('plugins.ccc')
+require('plugins.gitsigns')
+require('plugins.indentline')
+require('plugins.lualine')
+require('plugins.nvim-tree')(register, which.nvimtree)
+require('plugins.scrollbar')
+require('plugins.treesitter')
+require('plugins.virtcolumn')
