--- lua/plugins/interface/indentline.lua
@@ -1,2 +1,0 @@
--- see syntax.lua for IndentMarker
-require("ibl").setup({ indent = { highlight = { 'IndentMarker' } , char = '│' }})
