+++ lua/plugins/nvim-tree.lua
